==> 1.Pre_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os
import gc
import time
from tqdm import tqdm
import sys
import requests
import json
import re
import warnings
warnings.filterwarnings('ignore')

app_data = pd.read_csv('../data/app.csv')
sample_data = pd.read_csv('../data/sample.csv')
test_data = pd.read_csv('../data/test.csv')
train_data = pd.read_csv('../data/train.csv')
user_data = pd.read_csv('../data/user.csv')

pickle_path = "../pickle"
if not os.path.exists(pickle_path):
    os.mkdir(pickle_path)

# 处理app_data

'''
app.csv包含114584条数据，
2个字段：deviceid，applist
deviceid：用户设备id 
applist：用户所拥有的app
我们已将app的名字设置成了app_1,app_2..的形式。
'''
app_data.head()

app_data.isnull().any()

def flatten_app(df):    
    u = []
    a = []
    for i in tqdm(range(len(df['deviceid'].values))):
        u += [df['deviceid'].values[i]]*df['app_len'].values[i]
        a += list(df['applist'].values[i])
        
    new_df = pd.DataFrame()
    new_df['deviceid'] = u
    new_df['applist'] = a
        
    return new_df

new_app_data = app_data.groupby(by='deviceid').sum() 

new_app_data.head()

new_app_data['applist'] = new_app_data['applist'].map(lambda x:x.replace('][','').replace('[','').replace(' ]','').split(' '))
new_app_data['app_len'] =  new_app_data['applist'].map(lambda x:len(x))
new_app_data.reset_index(inplace = True)
deal_new_app_data = flatten_app(new_app_data)

new_app_data['deviceid'].unique().shape

new_app_data.shape

deal_new_app_data

if not os.path.exists("{}/device_new_app.pickle".format(pickle_path)):
    t1 = time.time()
    new_app_data.to_pickle("{}/device_new_app.pickle".format(pickle_path))
    print('ACTIVE TO PICKLE: ',time.time()-t1)

if not os.path.exists("{}/deal_device_new_app.pickle".format(pickle_path)):
    t1 = time.time()
    deal_new_app_data.to_pickle("{}/deal_device_new_app.pickle".format(pickle_path))
    print('ACTIVE TO PICKLE: ',time.time()-t1)

# sample_data 提交示例

'''
sample.csv为提交答案的格式，
id列和test.csv中的id列是对应关系，
target列为您预测的该视频是否会被点击，
1为点击0为未点击。 
'''
sample_data.head()

# ｔｅｓｔ　　ｄａｔａ

'''
test.csv包含3653592条数据
13个字段：id,deviceid,newsid,guid,
        pos,app_version,device_vendor,
        netmodel,osversion,lng,lat,
        device_version,ts 
id：代表数据集的第几条数据，从test_1到test_3653592。 
deviceid：用户的设备id。 
newsid：视频的id。 
guid：用户的注册id。 
pos：视频推荐位置。 
app_version：app版本。 
device_vendor：设备厂商。 
netmodel：网络类型。 
osversion：操作系统版本。 
lng：经度。 
lat：维度。 
device_version：设备版本。 
ts：视频暴光给用户的时间戳。
'''
test_data.head()

test_data['deviceid'].unique().shape

test_data['newsid'].unique().shape

if not os.path.exists("{}/test_data.pickle".format(pickle_path)):
    t1 = time.time()
    test_data.to_pickle("{}/test_data.pickle".format(pickle_path))
    print('ACTIVE TO PICKLE: ',time.time()-t1)

# train ｄａｔａ

'''
相比ｔｅｓｔ，多了两列特征target，timestamp
'''
train_data.head()

train_data.shape

train_data['deviceid'].unique().shape

train_data['newsid'].unique().shape

train_data['guid'].unique().shape

if not os.path.exists("{}/train_data.pickle".format(pickle_path)):
    t1 = time.time()
    train_data.to_pickle("{}/train_data.pickle".format(pickle_path))
    print('ACTIVE TO PICKLE: ',time.time()-t1)

# 处理user数据

'''
user.csv包含128573条数据，
9个字段：deviceid,guid,outertag,tag,
        level,personidentification,
        followscore,personalscore,
        gender 
deviceid：用户设备id。
guid：用户注册id。 
outertag：用户画像用|分隔，冒号后面的数字代表对该标签的符合程度，分数越高代表该标签越符合该用户。 
tag：同outertag。 
level：用户等级。 
personidentification：1表示劣质用户 0表示正常用户。 
followscore：徒弟分（好友分）。 
personalscore：个人分。 
gender：性别。 
'''
user_data.head(5)

user_data.isnull().any()

user_data['level'].unique()

user_data['level'] = user_data['level'].fillna(0)

user_data['followscore'].unique()

user_data['followscore'] = user_data['followscore'].fillna(0)

user_data['personalscore'].unique()

user_data['guid'].isnull().sum() / user_data['guid'].shape[0]

user_data['gender'].unique()

user_data['gender'].value_counts()

user_data['gender'].isnull().sum() / user_data['gender'].shape[0]

user_data['gender'] = user_data['gender'].fillna(-1)

user_data.isnull().any()

if not os.path.exists("{}/user_data.pickle".format(pickle_path)):
    t1 = time.time()
    user_data.to_pickle("{}/user_data.pickle".format(pickle_path))
    print('ACTIVE TO PICKLE: ',time.time()-t1)

tag_data = user_data[['deviceid','outertag','tag']]

tag_data['new_tag'] = tag_data['outertag'].apply(lambda x: str(x).split('|') if x is not np.nan else x)

tag_data['new_tag_2'] = tag_data['tag'].apply(lambda x: str(x).split('|') if x is not np.nan else x)

tag_data

tag_data['all_tag'] = tag_data['new_tag'].astype(str) +','+ tag_data['new_tag_2'].astype(str) 
tag_data.replace('nan,nan',np.nan,inplace = True)
tag_data['all_tag'] = tag_data['all_tag'].map(lambda x:str(x).replace('nan,[','[').replace('],nan',']').replace('],[',', '))

tag_data.loc[2,'all_tag']

tag_data['all_tag_word'] = tag_data['all_tag'].map(lambda x: re.sub("[A-Za-z0-9\!\%\[\]\,\。\_\:\.]", "", x))
tag_data['all_tag_word']= tag_data['all_tag_word'].replace("",'未知')
tag_data['all_tag_word'] = tag_data['all_tag_word'].map(lambda x:x.replace("'","").split(' '))

tag_data['all_tag_weight'] = tag_data['all_tag'].map(lambda x: re.findall("\d+\.?\d*", x))
tag_data['all_tag_weight']= tag_data['all_tag_weight'].map(lambda x: [0] if len(x) == 0 else x)

# tag_data['all_tag_weight']= tag_data['all_tag_weight'].replace("",0)

tag_data = tag_data[['deviceid','all_tag_word','all_tag_weight']]

tag_data.loc[21,'all_tag_word']

tag_data.loc[21,'all_tag_weight']

tag_data = tag_data.groupby(by='deviceid').sum() 

tag_data.reset_index(inplace = True)

tag_data.head()

tag_data['all_tag_word'] = tag_data['all_tag_word'].map(lambda x:[i for i in x if i != ''])

tag_data['all_tag_weight'] = tag_data['all_tag_weight'].map(lambda x:[i for i in x if i != ''])

tag_data.shape

tag_data['deviceid'].unique().shape

tag_data[['deviceid','all_tag_word','all_tag_weight']]

if not os.path.exists("{}/tag_weight_data.pickle".format(pickle_path)):
    t1 = time.time()
    tag_data[['deviceid','all_tag_word','all_tag_weight']].to_pickle("{}/tag_weight_new_data.pickle".format(pickle_path))
    print('ACTIVE TO PICKLE: ',time.time()-t1)

def flatten_app(df):    
    u = []
    a = []
    for i in tqdm(range(len(df['deviceid'].values))):
        u += [df['deviceid'].values[i]]*df['tag_len'].values[i]
        a += list(df['all_tag_word'].values[i])
        
    new_df = pd.DataFrame()
    new_df['deviceid'] = u
    new_df['all_tag_word'] = a
        
    return new_df

tag_data['tag_len'] = tag_data['all_tag_word'].map(lambda x:len(x))

deal_tag_data = flatten_app(tag_data)

deal_tag_data.shape

deal_tag_data.drop_duplicates(['deviceid','all_tag_word'],inplace = True)

deal_tag_data.shape

if not os.path.exists("{}/del_tag_data.pickle".format(pickle_path)):
    t1 = time.time()
    deal_tag_data.to_pickle("{}/deal_tag_new_data.pickle".format(pickle_path))
    print('ACTIVE TO PICKLE: ',time.time()-t1)
