==> 2. Graph Feature.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import lightgbm as lgb
import xgboost as xgb
from sklearn.preprocessing import LabelEncoder,OneHotEncoder
from sklearn.feature_extraction.text import CountVectorizer,TfidfVectorizer,HashingVectorizer
from sklearn.decomposition import TruncatedSVD,SparsePCA
from sklearn.model_selection import KFold,StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score,roc_auc_score,f1_score,recall_score

import gc
import time
import os
import sys
import warnings
warnings.filterwarnings('ignore')

pickle_path = "../pickle/"
deal_device_app = pd.read_pickle("{}/deal_device_new_app.pickle".format(pickle_path))
deal_tag_data = pd.read_pickle("{}/deal_tag_new_data.pickle".format(pickle_path))
train_data = pd.read_pickle("{}/train_data.pickle".format(pickle_path))
test_data = pd.read_pickle("{}/test_data.pickle".format(pickle_path))
deal_device_app.shape,train_data.shape,test_data.shape

deal_device_app.head()

deal_tag_data.head()

# Graph Feature
import scipy.sparse
from scipy import linalg
from scipy.special import iv
import scipy.sparse as sp

from sklearn import preprocessing
from sklearn.utils.extmath import randomized_svd
from sklearn.decomposition import TruncatedSVD

import argparse
import time

class ProNE():
    def __init__(self, G, emb_size=128, step=10, theta=0.5, mu=0.2, n_iter=5, random_state=2019):
        self.G = G
        self.emb_size = emb_size
        self.G = self.G.to_undirected()
        self.node_number = self.G.number_of_nodes()
        self.random_state = random_state
        self.step = step
        self.theta = theta
        self.mu = mu
        self.n_iter = n_iter
        
        mat = scipy.sparse.lil_matrix((self.node_number, self.node_number))
        print(mat.shape)
        for e in tqdm(self.G.edges()):
            if e[0] != e[1]:
                mat[int(e[0]), int(e[1])] = 1
                mat[int(e[1]), int(e[0])] = 1
        self.mat = scipy.sparse.csr_matrix(mat)
        print(mat.shape)

    def get_embedding_rand(self, matrix):
        # Sparse randomized tSVD for fast embedding
        t1 = time.time()
        l = matrix.shape[0]
        smat = scipy.sparse.csc_matrix(matrix)  # convert to sparse CSC format
        print('svd sparse', smat.data.shape[0] * 1.0 / l ** 2)
        U, Sigma, VT = randomized_svd(smat, n_components=self.emb_size, n_iter=self.n_iter, random_state=self.random_state)
        U = U * np.sqrt(Sigma)
        U = preprocessing.normalize(U, "l2")
        print('sparsesvd time', time.time() - t1)
        return U

    def get_embedding_dense(self, matrix, emb_size):
        # get dense embedding via SVD
        t1 = time.time()
        U, s, Vh = linalg.svd(matrix, full_matrices=False, check_finite=False, overwrite_a=True)
        U = np.array(U)
        U = U[:, :emb_size]
        s = s[:emb_size]
        s = np.sqrt(s)
        U = U * s
        U = preprocessing.normalize(U, "l2")
        print('densesvd time', time.time() - t1)
        return U

    def fit(self, tran, mask):
        # Network Embedding as Sparse Matrix Factorization
        t1 = time.time()
        l1 = 0.75
        C1 = preprocessing.normalize(tran, "l1")
        neg = np.array(C1.sum(axis=0))[0] ** l1

        neg = neg / neg.sum()

        neg = scipy.sparse.diags(neg, format="csr")
        neg = mask.dot(neg)
        print("neg", time.time() - t1)

        C1.data[C1.data <= 0] = 1
        neg.data[neg.data <= 0] = 1

        C1.data = np.log(C1.data)
        neg.data = np.log(neg.data)

        C1 -= neg
        F = C1
        features_matrix = self.get_embedding_rand(F)
        return features_matrix

    def chebyshev_gaussian(self, A, a, order=10, mu=0.5, s=0.5):
        # NE Enhancement via Spectral Propagation
        print('Chebyshev Series -----------------')
        t1 = time.time()

        if order == 1:
            return a

        A = sp.eye(self.node_number) + A
        DA = preprocessing.normalize(A, norm='l1')
        L = sp.eye(self.node_number) - DA

        M = L - mu * sp.eye(self.node_number)

        Lx0 = a
        Lx1 = M.dot(a)
        Lx1 = 0.5 * M.dot(Lx1) - a

        conv = iv(0, s) * Lx0
        conv -= 2 * iv(1, s) * Lx1
        for i in range(2, order):
            Lx2 = M.dot(Lx1)
            Lx2 = (M.dot(Lx2) - 2 * Lx1) - Lx0
            #         Lx2 = 2*L.dot(Lx1) - Lx0
            if i % 2 == 0:
                conv += 2 * iv(i, s) * Lx2
            else:
                conv -= 2 * iv(i, s) * Lx2
            Lx0 = Lx1
            Lx1 = Lx2
            del Lx2
            print('Bessell time', i, time.time() - t1)
        mm = A.dot(a - conv)
        self.embeddings = self.get_embedding_dense(mm, self.emb_size)
        return self.embeddings
    
    def transform(self):
        if self.embeddings is None:
            print("Embedding is not train")
            return {}
        self.embeddings = pd.DataFrame(self.embeddings)
        self.embeddings.columns = ['ProNE_Emb_deviceid_{}'.format(i) for i in range(len(self.embeddings.columns))]
        self.embeddings = self.embeddings.reset_index().rename(columns={'index' : 'nodes'}).sort_values(by=['nodes'],ascending=True).reset_index(drop=True)

        return self.embeddings

from tqdm import tqdm
import networkx as nx
import igraph as ig

def get_graph_embedding(df,prefix,fea1,fea2,size = 32,slow_fea = False):
    
    fea1_lbl,fea2_lbl = LabelEncoder(),LabelEncoder()
    new_fea1 = 'new_' + fea1
    new_fea2 = 'new_' + fea2
    df[new_fea1] = fea1_lbl.fit_transform(df[fea1])
#     deal_device_app.sort_values(by = new_fea1,inplace=True)
    df[new_fea2] = fea2_lbl.fit_transform(df[fea2])
    df[new_fea2] += df[new_fea1].max() + 1
    #df.drop_duplicates([new_fea1,new_fea2],inplace = True)
    
    print("Encoder Finished...")
    
    G = ig.Graph(directed = True)
    G.add_vertices(df[new_fea2].max()+1)
    G.add_edges(df[[new_fea1,new_fea2]].values)
    print(df[[new_fea1,new_fea2]].head())
    print("Build Graph Finished...")
    evcent = G.evcent() # 计算图中节点的向量中心性
    shell_index = G.shell_index() # 计算图中节点度至少为K的最大子图
    degree = G.degree() # 总度数
    pagerank = G.pagerank() # pagerank
    
    G_stat = pd.DataFrame()
    
    # 以下4个计算的很慢，效果还不错，可以注释掉，观察evcent的效果
    if(slow_fea):
        closeness = G.closeness() # 计算节点与网络中其他所有节点的距离的平均值   
        betweenness = G.betweenness() # 计算节点的介值
        constraint = G.constraint()
        eccentricity = G.eccentricity() # 计算给定节点到图中其他节点的最短距离的最大值。
    
    
        G_stat['closeness'] = closeness
        G_stat['betweenness'] = betweenness
        G_stat['constraint'] = constraint
        G_stat['eccentricity'] = eccentricity
   
    G_stat['evcent'] = evcent
    G_stat['shell_index'] = shell_index
    G_stat['degree'] = degree
    G_stat['pagerank'] = pagerank
    print("PR Finished...")
    
    G_stat = G_stat.reset_index()
    G_stat = G_stat[G_stat['index'].isin(df[new_fea1])]
    G_stat['index'] = fea1_lbl.inverse_transform(G_stat['index'])
    print(G_stat.head())
    print(G_stat.shape)
    print("Graph Stat Finished...")
    
    G_stat.to_csv("{}/Graph_edges_{}.csv".format(pickle_path,prefix),index = False)
    
    del G
    
    import gc
    gc.collect()
    
    G = nx.Graph()
    G.add_edges_from(df[[new_fea1,new_fea2]].values)
    model = ProNE(G,emb_size=size,n_iter=6,step=12)
    features_matrix = model.fit(model.mat, model.mat)
    model.chebyshev_gaussian(model.mat, features_matrix, model.step, model.mu, model.theta)
    emb = model.transform()
    fea = emb[emb['nodes'].isin(df[new_fea1])]
    fea['nodes'] = fea1_lbl.inverse_transform(fea['nodes'])
    fea.rename(columns={'nodes' : fea1},inplace=True)
    del G
    gc.collect()
    print("Embedding Finished...")
    print(fea.head())
    fea.to_csv("{}/Graph_prone_{}.csv".format(pickle_path,prefix),index = False)
    
    return fea,G_stat

fea_tag,stat_tag = get_graph_embedding(deal_tag_data,'tag','deviceid','all_tag_word',size = 32)

fea_app,stat_app = get_graph_embedding(deal_device_app,'app','deviceid','applist',size = 32)

train_data.shape,test_data.shape

all_data = pd.concat([train_data,test_data],sort = False,axis = 0)
all_data.shape

all_data.head()

#自己电脑内存不够，需要用服务器跑
fea_train_device_news,stat_train_device_news = get_graph_embedding(all_data,'all_data_dev_news','deviceid','newsid',size = 48)
