==> 3. Word Vector & Sparse Matrix.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import lightgbm as lgb
import xgboost as xgb
from sklearn.preprocessing import LabelEncoder,OneHotEncoder
from sklearn.feature_extraction.text import CountVectorizer,TfidfVectorizer,HashingVectorizer
from sklearn.decomposition import TruncatedSVD,SparsePCA
from sklearn.model_selection import KFold,StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score,roc_auc_score,f1_score,recall_score

import gc
import time
import os
import sys
import warnings
warnings.filterwarnings('ignore')

pickle_path = "../pickle"
vector_path = "../vector"

app_data_nlp = pd.read_pickle("{}/device_new_app.pickle".format(pickle_path))
user_fav_nlp = pd.read_pickle("{}/tag_weight_new_data.pickle".format(pickle_path))


app_data_nlp.head()

user_fav_nlp.head()

#glove model
from glove import *


t1 = time.time()
c = Corpus()
c.fit(app_data_nlp['applist'].values)
glove = Glove(no_components=300, learning_rate=0.05) 
glove.fit(c.matrix,epochs=12,no_threads=30,verbose=1)
glove.add_dictionary(c.dictionary)
glove.save("{}/app_data_glove300.model".format(vector_path))
print(time.time()-t1)
# 耗时51.98334240913391


t1 = time.time()
c = Corpus()
c.fit(user_fav_nlp['all_tag_word'].values)
glove = Glove(no_components=300, learning_rate=0.05) 
glove.fit(c.matrix,epochs=12,no_threads=30,verbose=1)
glove.add_dictionary(c.dictionary)
glove.save("{}/user_favorite_glove300.model".format(vector_path))
print(time.time()-t1)
# 耗时179.15632033348083

# Gen W2V Vector
from gensim import models


t1 = time.time()
w2v = models.Word2Vec(app_data_nlp['applist'].values, size=300, window=20, workers=40,hs=1) # 设置sg的话 变成skip-gram方法 我们测试效果差不多
w2v.wv.save_word2vec_format("{}/app_data_w2v300.model".format(vector_path))
print(time.time()-t1)
# 耗时19.42617154121399


t1 = time.time()
w2v = models.Word2Vec(user_fav_nlp['all_tag_word'].values, size=300, window=20, workers=40,hs=1) # 设置sg的话 变成skip-gram方法 我们测试效果差不多
w2v.wv.save_word2vec_format("{}/user_fav_w2v300.model".format(vector_path))
print(time.time()-t1)
# 耗时43.82200646400452

t = []
c = []

t1 = time.time()
tfidf = TfidfVectorizer(analyzer='word',token_pattern=u"(?u)\\b\\w+\\b",min_df=1,ngram_range=(1,1))
t.append(tfidf.fit_transform(app_data_nlp['applist'].map(lambda x:' '.join(x)).values))
cv = CountVectorizer(analyzer='word',token_pattern=u"(?u)\\b\\w+\\b",min_df=1,ngram_range=(1,1))
c.append(cv.fit_transform(app_data_nlp['applist'].map(lambda x:' '.join(x)).values))  
print(time.time()-t1)

t1 = time.time()
tfidf = TfidfVectorizer(analyzer='word',token_pattern=u"(?u)\\b\\w+\\b",min_df=1,ngram_range=(1,1))
t.append(tfidf.fit_transform(user_fav_nlp['all_tag_word'].map(lambda x:' '.join(x)).values))
cv = CountVectorizer(analyzer='word',token_pattern=u"(?u)\\b\\w+\\b",min_df=1,ngram_range=(1,1))
c.append(cv.fit_transform(user_fav_nlp['all_tag_word'].map(lambda x:' '.join(x)).values))
print(time.time()-t1)

from scipy import sparse
if not os.path.exists("{}/Sparse_Matrix".format(vector_path)):
    os.mkdir("{}/Sparse_Matrix".format(vector_path))
sparse.save_npz('{}/Sparse_Matrix/app_data_tfidf.npz'.format(vector_path), t[0])
sparse.save_npz('{}/Sparse_Matrix/user_fav_tfidf.npz'.format(vector_path), t[1])

sparse.save_npz('{}/Sparse_Matrix/app_data_count.npz'.format(vector_path), c[0])
sparse.save_npz('{}/Sparse_Matrix/user_fav_count.npz'.format(vector_path), c[1])

t[0]

c[0]
