==> 5. Flatten Stat Feature.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import lightgbm as lgb
import xgboost as xgb
from sklearn.preprocessing import LabelEncoder,OneHotEncoder
from sklearn.feature_extraction.text import CountVectorizer,TfidfVectorizer,HashingVectorizer
from sklearn.decomposition import TruncatedSVD,SparsePCA
from sklearn.model_selection import KFold,StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score,roc_auc_score,f1_score,recall_score

import gc
import time
import os
import sys
import warnings
warnings.filterwarnings('ignore')

pickle_path = "../pickle"
deal_device_new_app = pd.read_pickle("{}/deal_device_new_app.pickle".format(pickle_path))
device_new_app = pd.read_pickle("{}/device_new_app.pickle".format(pickle_path))

deal_device_new_app.head()

device_new_app.head()

from tqdm import tqdm

from collections import Counter

def Gini(pr):
    cate=Counter(pr)
    gini=1
    total=len(pr)
    for i in cate:
        gini=gini-(cate[i]/total)**2
    return gini

import math
# Calc Entropy
def entropy(pr):
    cate = Counter(pr)
    log2 = math.log2
    total = len(pr)
    ent = 0
    for i in cate:
        p = float(cate[i] / total)
        if p==0:
            ent=0
            continue
        ent = ent - p * (log2(p))
    return ent

def get_small(x,a,b):
    return int((x>a) & (x<b))

def get_feature_flatten(df):
    
    fea = []
    # Count Encoder
    t1 = time.time()
    df['appid_count'] = df.groupby(['applist'])['deviceid'].transform('count')
#     print(df.isnull().any())
#     print(df[['deviceid','appid_count']].groupby(['deviceid'])['appid_count'].transform('count').value_counts())
    fea.append(df[['deviceid','appid_count']].groupby(['deviceid'])['appid_count'].agg({'mean','std','min','max','median'}))
    fea.append(pd.DataFrame(df[['deviceid','applist']].groupby(['deviceid'])['applist'].apply(Gini)).rename(columns = {'applist':"Gini"}))
    fea.append(pd.DataFrame(df[['deviceid','applist']].groupby(['deviceid'])['applist'].apply(entropy)).rename(columns = {'applist':"entropy"}))
    df['appid_count_0_1e3'] = df['appid_count'].map(lambda x:get_small(x,0,1e3))
    df['appid_count_1e3_1e4'] = df['appid_count'].map(lambda x:get_small(x,1e3,1e4))
    df['appid_count_1e4_2e5'] = df['appid_count'].map(lambda x:get_small(x,1e4,2e5))
    tmp = df.groupby(['deviceid']).agg({'appid_count_0_1e3' : ['mean','sum','std'],
                          'appid_count_1e3_1e4' : ['mean','sum','std'],
                          'appid_count_1e4_2e5' : ['mean','sum','std']
                          })

    tmp.columns = ['_'.join(col).strip() for col in tmp.columns.values]
    fea.append(tmp)
    print("Count Active Finish... :",time.time()-t1)
    
    return fea

deviceid = device_new_app[['deviceid']]

deviceid.shape

fea0 = get_feature_flatten(deal_device_app)

for i in tqdm(fea0):
    print("~~~~~~~~~~~~~~~~~~~~~~~~~~~~")
    print(i.columns)

for i in tqdm(fea0):
    deviceid = deviceid.merge(i,how='left',on='deviceid')

deviceid.head()

deviceid.shape

deviceid = deviceid.set_index('deviceid').add_prefix("app_").reset_index()

deviceid.head()

deviceid.to_pickle("../pickle/app_flatten_all_data.pickle")
