==> 6. W2V Feature.ipynb <==
# 词向量与主题词挖掘

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import lightgbm as lgb
import xgboost as xgb
from sklearn.preprocessing import LabelEncoder,OneHotEncoder
from sklearn.feature_extraction.text import CountVectorizer,TfidfVectorizer,HashingVectorizer
from sklearn.decomposition import TruncatedSVD,SparsePCA
from sklearn.model_selection import KFold,StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score,roc_auc_score,f1_score,recall_score
import datetime
import gc
import time
import os
import sys
import warnings
warnings.filterwarnings('ignore')

pickle_path = "../pickle"

device_app = pd.read_pickle("{}/device_new_app.pickle".format(pickle_path))
tag_data = pd.read_pickle("{}/tag_weight_new_data.pickle".format(pickle_path))

device_app.shape,tag_data.shape

device_app.head()

tag_data.head()

from gensim import corpora, models, similarities
from gensim.models.doc2vec import TaggedDocument
from glove import *


def get_gensim_feature(now=None,model='word2vec',fea = 'fea_name',size=5,window=10,prefix='active'):
    begin=datetime.datetime.now()
    print("===================================================================")
    print(prefix,model,'begin')
    df = now.copy()
    if os.path.exists("../pickle/{}_{}_emb.pickle".format(prefix,model)):
        return pd.read_pickle("../pickle/{}_{}_emb.pickle".format(prefix,model))
    else:
        dictionary = corpora.Dictionary(df[fea].values)
        corpus = [dictionary.doc2bow(text) for text in df[fea].values]
        if model=='word2vec':
            '''
            word2vec:
            将所有单词映射为向量，如果一个人有多个标签，
            每个标签代表一个词，则吧每个标签映射成的单词向量相加然后取平均
            
            相当于把每一个ｄｅｖｉｃｅｉｄ对应的ａｐｐ，ｔａｇ，映射为一个单词（同维度）
            '''
            if os.path.exists("../vector/w2v_{}.model".format(prefix)):
                w2v = models.KeyedVectors.load_word2vec_format("../vector/w2v_{}.model".format(prefix), binary=False)
            else:
                w2v = models.Word2Vec(df[fea].values, size=size, window=window, workers=40)
                w2v.wv.save_word2vec_format("../vector/w2v_{}.model".format(prefix))
            vocab = list(w2v.wv.vocab.keys())

            # Sentence Embedding

            w2v_feature = np.zeros((df.shape[0],size))
            w2v_feature_avg = np.zeros((df.shape[0],size))

            for i,line in tqdm(enumerate(df[fea].values.tolist())):
                num = 0
                if line == '':
                    w2v_feature_avg[i,:] = np.zeros(size)
                else:
                    for word in line:
                        num += 1
                        vec = w2v[word] if word in vocab else np.zeros(size)
                        w2v_feature[i,:] += vec
                    w2v_feature_avg[i,:] = w2v_feature[i,:] / num
            w2v_avg = pd.DataFrame(w2v_feature_avg)
            w2v_avg = w2v_avg.add_prefix("W2V_AVG_{}_".format(prefix))
            w2v_avg['deviceid'] = df['deviceid']
            df_agg = w2v_avg

        elif model=='lda':
            '''
            LDA文档主题生成模型，也称三层贝叶斯概率模型，包含词、主题和文档三层结构。
            
            本程序是提取出每一行语料的２０个主题词
            col　每一行代表数据中的每一行，有２０个主题词
            '''
            lda = models.LdaModel(corpus=corpus, id2word=dictionary, num_topics=size)
            col = np.zeros((df.shape[0],20))
            ans = lda.get_document_topics(corpus)
            for i in tqdm(range(df.shape[0])):
                for j in ans[i]:
                    col[i][j[0]] = j[1]

            df_agg = pd.DataFrame(col)
            df_agg = df_agg.add_prefix("LDA_TOPIC_{}_".format(prefix))
            df_agg['deviceid'] = df['deviceid']

        elif model=='fasttext':
            
            '''
            fasttext:与word2vec，都是将单词映射为向量
            将所有单词映射为向量，如果一个人有多个标签，
            每个标签代表一个词，则吧每个标签映射成的单词向量相加然后取平均
            
            相当于把每一个ｄｅｖｉｃｅｉｄ对应的ａｐｐ，ｔａｇ，映射为一个单词（同维度）
            
            '''
            
            if os.path.exists("../vector/fasttext_{}.model".format(prefix)):
                fasttext = models.KeyedVectors.load_word2vec_format("../vector/fasttext_{}.model".format(prefix), binary=False)
            else:
                fasttext = models.FastText(df[fea].values, size=size, window=window, workers=40)
                fasttext.wv.save_word2vec_format("../vector/fasttext_{}.model".format(prefix))
            vocab = list(fasttext.wv.vocab.keys())

            fasttext_feature = np.zeros((df.shape[0],size))
            fasttext_feature_avg = np.zeros((df.shape[0],size))

            for i,line in tqdm(enumerate(df[fea].values.tolist())):
                num = 0
                if line == '':
                    fasttext_feature_avg[i,:] = np.zeros(size)
                else:
                    for word in line:
                        num += 1
                        vec = fasttext[word] if word in vocab else np.zeros(size)
                        fasttext_feature[i,:] += vec
                    fasttext_feature_avg[i,:] = fasttext_feature[i,:] / num
            fasttext_avg = pd.DataFrame(fasttext_feature_avg)
            fasttext_avg = fasttext_avg.add_prefix("FASTTEXT_AVG_{}".format(prefix))
            fasttext_avg['deviceid'] = df['deviceid']
            df_agg = fasttext_avg

        elif model=='doc2vec':
            '''
            Doc2vec是一个非常好的技术。它易于使用，效果很好，而且从名称上可以理解，很大程度上基于word2vec
            word2vec是将一个单词映射为向量
            Doc2vec是将一句话或者一片文档映射为向量
            '''
            if os.path.exists("../vector/d2v_{}.model".format(prefix)):
                d2v = models.KeyedVectors.load_word2vec_format("../vector/d2v_{}.model".format(prefix), binary=False)
            else:            
                docs = [TaggedDocument(words=i[1],tags=[str(i[0])]) for i in df[['deviceid',fea]].values]
                d2v = models.Doc2Vec(docs,size=size,window=window,workers=40)
                d2v.wv.save_word2vec_format("../vector/d2v_{}.model".format(prefix))
            vocab = list(d2v.wv.vocab.keys())
            
            d2v_avg = np.zeros((df.shape[0],size))
            
            for index,i in tqdm(enumerate(df[fea].values)):
                line = []
                for j in i:
                    line.append(d2v[j] if j in vocab else 0)
                d2v_avg[index,] = np.mean(line,axis=0)
            
            d2v_avg = pd.DataFrame(d2v_avg)
            d2v_avg = d2v_avg.add_prefix("d2v_AVG_{}".format(prefix))
            d2v_avg['deviceid'] = df['deviceid']
            df_agg = d2v_avg

        elif model=='lsi':
            '''
            主题模型，将每一行样本所对应的主题提取出来
            '''
            lsi = models.LsiModel(corpus=corpus, id2word=dictionary, num_topics=size)
            df_agg = []
            for i in tqdm(df[fea].values):
                lsi_ = lsi[dictionary.doc2bow(i)]
                df_agg.append([tmp[1] for tmp in lsi[lsi_]])

            df_agg = pd.DataFrame(df_agg)
            df_agg = df_agg.add_prefix("LSI_TOPIC_{}_".format(prefix))
            df_agg['deviceid'] = df['deviceid']
            
        elif model=='glove':
            '''
            词向量模型，将单词映射为向量
            '''
            matrix =  Corpus()
            matrix.fit(df[fea].values)
            glove = Glove(no_components=size, learning_rate=0.05)
            glove.fit(matrix.matrix,epochs=10,no_threads=30,verbose=1)
            glove.add_dictionary(matrix.dictionary)
            ans = np.zeros((df.shape[0],size))
            for index,i in tqdm(enumerate(df[fea].values)):
                line = []
                for j in i:
                    line.append(glove.word_vectors[glove.dictionary[j]])
                ans[index,] = np.mean(line,axis=0)
            df_agg = pd.DataFrame(ans)
            df_agg = df_agg.add_prefix("Glove_AVG_{}".format(prefix))
            df_agg['deviceid'] = df['deviceid']
            
        df_agg.to_pickle("../pickle/{}_{}_emb.pickle".format(prefix,model))
    end=datetime.datetime.now()
    dur=end-begin
    print(begin)
    print(end)
    print(dur)
    print(prefix,model,'end')
    print("===================================================================")
    return df_agg

from tqdm import tqdm
lsi = get_gensim_feature(device_app,'lsi','applist',32,10,'device_app')
w2v = get_gensim_feature(device_app,'word2vec','applist',64,10,'device_app')
fasttext = get_gensim_feature(device_app,'fasttext','applist',64,10,'device_app')
d2v = get_gensim_feature(device_app,'doc2vec','applist',64,10,'device_app')
lda = get_gensim_feature(device_app,'lda','applist',20,10,'device_app')
glove = get_gensim_feature(device_app,'glove','applist',20,10,'device_app')


lsi_1 = get_gensim_feature(tag_data,'lsi','all_tag_word',32,10,'tag_data')
w2v_1 = get_gensim_feature(tag_data,'word2vec','all_tag_word',64,10,'tag_data')
fasttext_1 = get_gensim_feature(tag_data,'fasttext','all_tag_word',64,10,'tag_data')
d2v_1 = get_gensim_feature(tag_data,'doc2vec','all_tag_word',64,10,'tag_data')
lda_1 = get_gensim_feature(tag_data,'lda','all_tag_word',20,10,'tag_data')
glove_1 = get_gensim_feature(tag_data,'glove','all_tag_word',20,10,'tag_word')


lsi.head()

w2v.head()

fasttext.head()

d2v.head()

lda.head()

glove.head()
